+++ spec/unit/ogr/coordinate_transformation_spec.rb
@@ -25,13 +25,15 @@ RSpec.describe OGR::CoordinateTransformation do
   describe '#initialize' do
     context 'source_srs is not an OGR::SpatialReference' do
       it 'raises an OGR::Failure' do
-        expect { described_class.new(123, dest_srs) }.to raise_exception(OGR::Failure)
+        expect { described_class.new(123, dest_srs) }
+          .to raise_exception GDAL::Error
       end
     end
 
     context 'dest_srs is not an OGR::SpatialReference' do
       it 'raises an OGR::Failure' do
-        expect { described_class.new(source_srs, 123) }.to raise_exception(OGR::Failure)
+        expect { described_class.new(source_srs, 123) }
+          .to raise_exception GDAL::Error
       end
     end
 
@@ -41,13 +43,6 @@ RSpec.describe OGR::CoordinateTransformation do
         expect(instance).to be_a described_class
         expect(instance.c_pointer).to be_a FFI::Pointer
       end
-    end
-  end
-
-  describe '#destroy!' do
-    it 'sets @c_pointer to nil' do
-      subject.destroy!
-      expect(subject.c_pointer).to be_nil
     end
   end